Fixed reflection refraction test error

+++ CSharp/Patterns/RingPattern.cs
@@ -11,7 +11,7 @@ namespace CSharp
 
         public override Color PatternAt(Point p)
         {
-            return Math.Floor(Math.Sqrt(p.x * p.x + p.z * p.z)) % 2 == 0 ? a : b;
+            return (int) Math.Floor(Math.Sqrt(p.x * p.x + p.z * p.z)) % 2 == 0 ? a : b;
         }
 
         protected override Color PatternAtShape(RTObject obj, Point localPoint)
